Move Prompts models
