+++ src/Auth0.ManagementApi/Clients/PromptsClient.cs
@@ -5,6 +5,7 @@ using System.IO;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
+using Auth0.ManagementApi.Models.Prompts;
 
 
 namespace Auth0.ManagementApi.Clients
