+++ src/Auth0.ManagementApi/Models/Prompts/PromptUpdateRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace Auth0.ManagementApi.Models
+﻿namespace Auth0.ManagementApi.Models.Prompts
 {
     /// <summary>
     /// Request configuration for updating prompt settings.
